dashboard/requirements.txt
psycopg2-binary==2.9.7
streamlit==1.27.0
streamlit-autorefresh==1.0.1
pandas==2.0.3